hw_2/pyproject.toml
[project]
name = "hw-2"
version = "0.1.0"
description = "homework 2"
authors = []
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "latex-generator-dazaii (>=0.1.0,<0.2.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
